+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Yii Mailer Library Change Log
 
-## 5.1.1 under development
+## 6.0.0 under development
 
-- no changes in this release.
+- New #84: Add `MessageInterface` methods: `getAttachments()`,  `getEmbeddings()`, `getHeaders()`,
+  `withAddedAttachments()`, `withAddedEmbeddings()` (@vjik)
+- Chg #84: Rename `MessageInterface` methods: `withEmbedded()` to `withEmbeddings()`, `withAttached()`
+  to `withAttachments()` and allow passing several files to them (@vjik)
+- New #84: Add `Message` class that implements `MessageInterface` (@vjik)
 
 ## 5.1.0 July 02, 2024